refactor(capture): move convenience functions back to plugin.capture
